Add files via upload
Project Objective:
The objective of this analysis is to discover various factors contributing to increased customer churn rate at the bank, and provide the business users with these insights which they can use to make informed decisions and strategize on how to improve customer retention and reduce churn rate.

Tools:
 I used Power BI and Power Query for this analysis and also do analysis by using the artificial neural network in jupyter notebook.

About the data
The raw data was provided in the form of a CSV file with 10,000 rows of data. Data was sourced from a commercial bank who was looking to improve its customer churn rate.